Added check for X-Content-Type-Options and refactored existing HTTP header check to use common base class.
